requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.6.0
Flask-Login==0.5.0
Flask-WTF==0.15.2
googlemaps==4.5.3
squareup==4.0.0
